+++ askanna/__init__.py
@@ -4,7 +4,7 @@
 
 __author__ = "AskAnna Team"
 __email__ = "devops@askanna.io"
-__version__ = "0.12.2"
+__version__ = "0.13.0"
 
 import re
 import sys