+++ askanna/core/utils.py
@@ -7,12 +7,14 @@ import os
 from pathlib import Path
 import re
 import sys
-from typing import Any, List, Callable, Tuple, Union, Dict
+from typing import Any, List, Tuple, Dict
 import uuid
 from zipfile import ZipFile
 
 import click
 import croniter
+from email_validator import validate_email, EmailNotValidError
+import igittigitt
 import pytz
 import tzlocal
 import requests
@@ -122,6 +124,10 @@ def object_fullname(o):
         return o.__class__.__name__  # Avoid reporting __builtin__
     else:
         return module + "." + o.__class__.__name__
+
+
+def flatten(t):
+    return [item for sublist in t for item in sublist]
 
 
 def update_available() -> bool:
@@ -279,29 +285,31 @@ def store_config(new_config):
 
 
 # Zip the files that matches the filter from given directory
-def zip_files_in_dir(
-    directory_path: str, zip_file: ZipFile, filter=lambda x: x
-) -> None:
-    files = get_files_in_dir(directory_path=directory_path, filter=filter)
+def zip_files_in_dir(directory_path: str, zip_file: ZipFile, ignore_file: str = None) -> None:
+    files = get_files_in_dir(directory_path=directory_path, ignore_file=ignore_file)
 
     # Iterate over all the files and zip them
     for file in sorted(files):
         zip_file.write(file)
 
 
-def get_files_in_dir(
-    directory_path: str, filter: Callable[[str], Union[str, bool]] = lambda x: x
-) -> set:
+def get_files_in_dir(directory_path: str, ignore_file: str = None) -> set:
     file_list = set()
+
+    ignore_parser = igittigitt.IgnoreParser()
+    if ignore_file:
+        ignore_parser.parse_rule_files(os.path.dirname(ignore_file), ignore_file)
 
     # Iterate over all the files in directory
     for root, _, files in os.walk(directory_path):
         for file in files:
-            if filter(file):
-                # Create complete filepath of file in directory
-                file_path = os.path.join(root, file)
-                if file_path.startswith("./"):
-                    file_path = file_path[2:]
+            # Create complete filepath of file in directory
+            file_path = os.path.join(root, file)
+            if file_path.startswith("./"):
+                file_path = file_path[2:]
+
+            # can we add this file to the collection?
+            if not ignore_parser.match(file_path):
                 file_list.add(file_path)
 
     return file_list
@@ -518,6 +526,7 @@ def validate_yml_job_names(config):
         "environment",
         "image",
         "job",
+        "notifications",
         "project",
         "push-target",
         "timezone",
@@ -584,22 +593,132 @@ def validate_yml_environments(config: Dict, jobname=None) -> bool:
     return True
 
 
-def validate_yml_job(config):
+def validate_yml_notifications(config: Dict, jobname=None) -> bool:
+    """
+    Validate the definition of `notifications` in global and job
+    """
+    notifications = config.get("notifications")
+    if notifications and not isinstance(notifications, dict):
+        if jobname:
+            click.echo(
+                f"Invalid definition of `notifications` found in job `{jobname}`:\n"
+                f"notifications: {notifications}"
+                "\n"
+                "For notifications documentation: https://docs.askanna.io/jobs/notifications/",
+                err=True,
+            )
+        else:
+            click.echo(
+                "Invalid `notifications` setting found:"
+                f"notifications: {notifications}"
+                "\n"
+                "For notifications documentation: https://docs.askanna.io/jobs/notifications/",
+                err=True,
+            )
+        return False
+    elif notifications and isinstance(notifications, dict):
+        # check for values of `all` and `error`, both should be dicts
+        noti_all = notifications.get("all")
+        noti_error = notifications.get("error")
+        if any(
+            [
+                noti_all and not isinstance(noti_all, dict),
+                noti_error and not isinstance(noti_error, dict),
+            ]
+        ):
+            if jobname:
+                click.echo(
+                    f"Invalid definition of `notifications` found in job `{jobname}`:\n"
+                    f"notifications: {notifications}"
+                    "\n"
+                    "For notifications documentation: https://docs.askanna.io/jobs/notifications/",
+                    err=True,
+                )
+            else:
+                click.echo(
+                    "Invalid `notifications` setting found:\n"
+                    f"notifications: {notifications}"
+                    "\n"
+                    "For notifications documentation: https://docs.askanna.io/jobs/notifications/",
+                    err=True,
+                )
+            return False
+
+        # proceed with validation of the content
+        # assume we can get values out of all/email and error/email
+        noti_all_values = notifications.get("all", {}).get("email", [])
+        noti_error_values = notifications.get("error", {}).get("email", [])
+
+        all_email = sorted(
+            flatten(
+                list(
+                    map(
+                        lambda x: x.split(","),
+                        noti_all_values + noti_error_values,
+                    )
+                )
+            )
+        )
+        for value in all_email:
+            if value.startswith("${") and value.endswith("}"):
+                # we found a variable not subsituted yet
+                continue
+            if value in ["workspace admins", "workspace members"]:
+                # valid addresees
+                continue
+
+            # validate on actual content of the value whether it is an e-mail address or not
+            try:
+                validate_email(value)
+            except EmailNotValidError:
+                click.echo(
+                    "Invalid `notifications/email` found:\n" f"{value}" "\n",
+                    err=True,
+                )
+                return False
+
+    return True
+
+
+def validate_askanna_yml(config):
+    """
+    Given an dictionary of askanna.yml validate each component
+    """
     jobs = config.items()
+
     global_timezone = config.get("timezone")
     # validate the global timezone
-    if global_timezone and global_timezone not in pytz.all_timezones:
-        click.echo(
-            "Invalid timezone setting found in askanna.yml:\n"
-            + f"   timezone: '{global_timezone}'",
-            err=True,
-        )
+    if global_timezone:
+        if global_timezone not in pytz.all_timezones:
+            click.echo(
+                "Invalid timezone setting found in askanna.yml:\n"
+                + f"timezone: {global_timezone}",
+                err=True,
+            )
+            return False
+        timezone_checked = True  # used later, so we only print a warning message about the timezone once
+    else:
+        timezone_checked = False  # used later, so we only print a warning message about the timezone once
+
+    # validate whether the global environment definitions are correct
+    if not validate_yml_environments(config):
+        return False
+
+    # validate for global notification settings
+    if not validate_yml_notifications(config):
+        return False
+
+    # validate jobs names
+    if not validate_yml_job_names(config):
         return False
 
     for jobname, job in jobs:
         if isinstance(job, dict):
             if not validate_yml_environments(job, jobname=jobname):
                 return False
+            if not validate_yml_notifications(job, jobname=jobname):
+                return False
+
             schedule = job.get("schedule")
             timezone = job.get("timezone")
             if not schedule:
@@ -613,12 +732,40 @@ def validate_yml_job(config):
                     )
                     return False
             # validate the timezone if set
+            timezone = job.get("timezone")
             if timezone and timezone not in pytz.all_timezones:
                 click.echo(
-                    f"Invalid timezone found in job: `{job}`: `{timezone}`",
+                    f"Invalid timezone setting found in job `{jobname}`:\n"
+                    + f"timezone: {timezone}",
                     err=True,
                 )
                 return False
+
+            # validate the schedule
+            schedule = job.get("schedule")
+            if schedule:
+                for cron_line, parsed in parse_cron_schedule(schedule):
+                    if not parsed:
+                        click.echo(
+                            f"Invalid schedule definition `{cron_line}` found in job `{jobname}`",
+                            err=True,
+                        )
+                        return False
+
+                if not timezone and not global_timezone and not timezone_checked:
+                    timezone_local = getLocalTimezone()
+                    if timezone_local != "UTC":
+                        click.echo(  # noqa
+                            f"""
+By default, the AskAnna platform uses time zone UTC. Your current time zone is {timezone_local}.
+To use your local time zone for scheduling jobs in this project, add the next line to your config in `askanna.yml`:
+
+timezone: {timezone_local}
+
+For more information, read the documentation: https://docs.askanna.io/jobs/create-job/#time-zone
+"""
+                        )
+                    timezone_checked = True
     return True